All parts scale

+++ MainWindow/Sleeve.cs
@@ -41,7 +41,7 @@ namespace KnitDesigner
                 throw new ArgumentException("Please make sure your gauge is a positve number.");
             }
             this.width = width * perInch;
-            return this.width;
+            return (decimal)this.width;
         }
 
     }
